backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
google-cloud-documentai==2.23.0
google-cloud-aiplatform==1.38.0
google-cloud-storage==2.10.0
google-cloud-firestore==2.13.0
google-cloud-dlp==3.12.0
google-cloud-translate==3.12.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
aiofiles==23.2.1
pillow==10.1.0
PyPDF2==3.0.1
python-docx==1.1.0
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
langchain==0.0.340
langchain-google-vertexai==0.0.3
chromadb==0.4.18
tiktoken==0.5.2
openai==1.3.7
anthropic==0.7.8
